+++ Assets/Gemserk.SelectionHistory/Editor/UnityBuiltInIcons.cs
@@ -1,0 +1,14 @@
+namespace Gemserk
+{
+    public static class UnityBuiltInIcons
+    {
+        public const string pickObjectIconName = "d_pick";
+        public const string favoriteIconName = "d_Favorite Icon";
+        public const string favoriteEmptyIconName = "d_Favorite";
+        public const string removeIconName = "d_ol_minus";
+        public const string tagIconName = "AssetLabelIcon";
+        public const string searchIconName = "d_Search Icon";
+        public const string eyeViewToolIconName = "d_ViewToolOrbit";
+        public const string refreshIconName = "TreeEditor.Refresh";
+    }
+}