+++ Assets/Gemserk.SelectionHistory/CHANGELOG.md
@@ -1,3 +1,26 @@
+# 1.1.3
+
+* Fixed bugs on scroll when selecting with new window version using UIElements.
+
+# 1.1.0
+
+* Separated Favorite assets in a new window and removed from history window.
+* There is now a Favorites asset to keep track of user's favorites, can be ignored or not from scm, depends on what the user want. 
+* Added icons for both windows title bar. 
+* Drag objects to fav window favorites them
+* Added some settings to selection history window custom menu. 
+* New history window using UIElements
+* Now user preferences for selection history can be open from custom menu.
+
+# 1.0.8
+
+* Now scene is pinged when pressing ping for unloaded objects.
+* Now all scene references are shown with scene name prefix.
+
+# 1.0.7
+
+* Feature: store selection from unloaded scenes.
+
 # 1.0.4
 
 * Fixed missing .meta was failing on reimporting package in readonly folder.